backend_py/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.2.0
numpy>=1.26.3
ultralytics>=8.1.0
opencv-python>=4.8.0
python-dotenv>=1.0.0
edge-tts>=6.1.9
openai>=1.0.0
httpx>=0.25.0
nvidia-riva-client
protobuf
googleapis-common-protos
